Squash in PROJ-150-implement-server-internal-game-structure (pull request #71)
PROJ-150 implement server internal game structure

* PROJ-151: Implemented create_player, update_player, and destroy_player

* PROJ-150 Created initial positions for players

* PROJ-150 Added a players_connected variable that tracks how many players are connected

* PROJ-151: Updated Game struct with float positions, Player** array, and update_game signature

* PROJ-151: Implemented create_game, destroy_game, add_player, and update_game

* PROJ-151: Changed pos_x and pos_y to float in Player struct

* PROJ-151: Changed pos_x and pos_y to float in Item struct

* PROJ-151: Changed pos_x and pos_y to float in packet structs

* PROJ-151: Removed unnecessary float casts in update_player

* PROJ-150 Made threads' flags static

* PROJ-150 Included necessary header file.

* PROJ-151: Fixed players_connected increment, bad cast in mail drop loop, removed unused void suppressions

* PROJ-150 Set Game status to NOT_READY when instantiating game

* PROJ-150 Removed unnecessary comments

* PROJ-150 Added functions to update game tick and game status

* PROJ-150 Added a new game status

* PROJ-151: Implemented game thread loop with state transitions and tick-based lifecycle

* PROJ-151: Separated connection loop and game loop in game thread

* PROJ-151: Refactor game_thread into separate connection and game loop functions

Split the monolithic run_game_t loop into run_connection_loop and run_game_loop static functions, leaving run_game_t as a clean entry point.

* PROJ-151: Fix type bugs and cleanup across game, player, and packet files

- Change vel_x/vel_y from int32_t/uint32_t to float in Player, ClientRegularPacket, and AuthPlayerRecord
- Remove unsafe casts in update_player and form_auth_packet
- Remove duplicate game->status assignment in create_game
- Log error on partial map position read in create_game
- Fix typo "nauthorized" in handle_init_packet log message

* PROJ-150 Removed termination of the game process in main

* PROJ-151: Add game integration test for connection and gameplay packet flow

Simulates PLAYERS_PER_MATCH clients: sends INIT packets, validates InitPacket
broadcast, then sends distinct regular packets per player each round and
validates AuthPacket records reflect the correct data across 5 consecutive
checks. Includes local Makefile and test-game target in top-level Makefile.

* PROJ-151: Fix test build path and validation timing in test_game

- Fix Makefile include paths (../../ not ../../../)
- Use memcmp for float field comparison to avoid -Wfloat-equal
- Fix Phase 3 validation: drain stale buffered packets before polling
  for the fresh AuthPacket to avoid off-by-one-round data mismatch

* PROJ-151: Add .gitignore for test/game obj dir, untrack compiled object

Approved-by: Daniel Pelley

+++ server/game/include/item.h
@@ -7,8 +7,8 @@ struct Item
 {
     uint8_t item_id;
 
-    uint32_t pos_x;
-    uint32_t pos_y;
+    float pos_x;
+    float pos_y;
 };
 
 #endif
